+++ java-client/src/main/java/io/sphere/internal/Categories.java
@@ -6,6 +6,13 @@ import io.sphere.client.QueryRequest;
 
 /** Sphere HTTP APIs for working with Categories in a given project. */
 public interface Categories {
-    /** Queries all categories. */
+    /**
+     * Queries categories. This method does not fetch all categories by default.
+     * @deprecated since 0.56.0, use {@link Categories#query() instead}
+     **/
+    @Deprecated
     QueryRequest<BackendCategory> all();
+
+    /** Queries all categories. */
+    QueryRequest<BackendCategory> query();
 }